requirements.txt
fastapi
uvicorn
boto3
python-multipart
Pillow
fastapi[all] 
starlette
mangum